[AsseticBundle] renamed some test files to respect the exclude of Resources in phpunit.xml.dist
