+++ src/Symfony/Bundle/AsseticBundle/Tests/FunctionalTest.php
@@ -11,7 +11,6 @@
 
 namespace Symfony\Bundle\AsseticBundle\Tests;
 
-use Symfony\Bundle\AsseticBundle\Tests\Kernel\TestKernel;
 use Symfony\Component\DomCrawler\Crawler;
 use Symfony\Component\HttpFoundation\Request;
 
@@ -23,17 +22,17 @@ class FunctionalTest extends \PHPUnit_Framework_TestCase
             $this->markTestSkipped('Assetic is not available.');
         }
 
-        $cache = __DIR__.'/Kernel/cache';
+        $cache = __DIR__.'/Resources/cache';
         if (!is_dir($cache)) {
             mkdir($cache);
         } else {
-            shell_exec('rm -rf '.escapeshellarg(__DIR__.'/Kernel/cache/*'));
+            shell_exec('rm -rf '.escapeshellarg(__DIR__.'/Resources/cache/*'));
         }
     }
 
     protected function tearDown()
     {
-        shell_exec('rm -rf '.escapeshellarg(__DIR__.'/Kernel/cache'));
+        shell_exec('rm -rf '.escapeshellarg(__DIR__.'/Resources/cache'));
     }
 
     /**
