+++ src/Symfony/Bundle/AsseticBundle/Tests/Resources/config/config.yml
