+++ src/Symfony/Bundle/AsseticBundle/Tests/Resources/config/routing.yml
