+++ src/Symfony/Bundle/AsseticBundle/Tests/Resources/views/layout.html.twig
