+++ src/Symfony/Bundle/AsseticBundle/Tests/Resources/web/javascript1.js
