+++ src/Symfony/Bundle/AsseticBundle/Tests/Resources/web/javascript2.js
