+++ src/Symfony/Bundle/AsseticBundle/Tests/Resources/web/stylesheet1.css
