+++ src/Symfony/Bundle/AsseticBundle/Tests/Resources/web/stylesheet2.css
